Add time module for simulation

+++ src/simulator/simulator.h
@@ -5,6 +5,7 @@
 #include "package/package.h"
 #include "priorityqueue/priorityqueue.h"
 #include "protocol/protocol.h"
+#include "time/time.h"
 #include <stdlib.h>
 
 /**
@@ -17,6 +18,6 @@
  * @param delay the mean delay to send a package
  * @param seed the random seed
  */
-void run_simulation(int messages, float corruption, float loss, float delay, int seed);
+void run_simulation(size_t messages, float corruption, float loss, float delay, int seed);
 
 #endif // SIMULATOR_H
